Don’t limit strip counts etc. to INT_MAX
* src/patch.c (debug): Now unsigned short int, not int, since
it is used as a mask and only the bottom 9 bits matter.
(patch_get, strippath, maxfuzz): Now intmax_t, not int.
(numeric_string): Return intmax_t, not int.
On overflow return an extremum rather than reporting an error,
since the resulting values are effectively infinity anyway.
All uses changed.
* src/util.c (success, fetchname, parse_name): Accept intmax_t, not int.

+++ src/common.h
@@ -95,7 +95,7 @@ extern struct outfile tmpin;
 extern struct outfile tmppat;
 
 #if DEBUGGING
-extern int debug;
+extern unsigned short int debug;
 #else
 # define debug 0
 #endif
@@ -105,9 +105,9 @@ extern bool noreverse_flag;
 extern bool reverse_flag;
 extern enum verbosity { DEFAULT_VERBOSITY, SILENT, VERBOSE } verbosity;
 extern bool skip_rest_of_patch;
-extern int strippath;
+extern intmax_t strippath;
 extern bool canonicalize_ws;
-extern int patch_get;
+extern intmax_t patch_get;
 extern bool set_time;
 extern bool set_utc;
 extern bool follow_symlinks;
